Fix EnemyCollision
Fix EnemyCollision - 현재 스킬 프리팹 자체에 skilldata가 안달려있음

+++ Assets/Scripts/Etc/SkillData.cs
@@ -8,6 +8,6 @@ public class SkillData : ScriptableObject
     public int skillLevel;
     public string skillName;
     public float skillCooldown;
-    public float skillDamage;
+    public int skillDamage;
     public Sprite skillSprite;
 }
